use int64_t for all num_elems or size calculations to support VERY large message sizes

+++ pccl/include/common.h
@@ -15,9 +15,9 @@
 } while(0)
 
 // Kernel for vector addition.
-__global__ void vectorAddKernel(float* a, const float* b, int n);
+__global__ void vectorAddKernel(float* a, const float* b, int64_t n);
 
 // Function to launch the kernel.
-void vectorAdd(float* a, const float* b, int n, cudaStream_t stream);
+void vectorAdd(float* a, const float* b, int64_t n, cudaStream_t stream);
 
 #endif // COMMON_H
